ml_engine/requirements.txt
# ======================
# ML Engine Dependencies
# ======================

# Core
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
imbalanced-learn==0.11.0

# Time Series
prophet==1.1.5
statsmodels==0.14.0

# Deep Learning
tensorflow==2.15.0
keras==3.0.0

# Computer Vision
pillow==10.1.0
opencv-python==4.8.1.78
albumentations==1.3.1

# Explainability
shap==0.44.0
lime==0.2.0.1

# Utilities
joblib==1.3.2
tqdm==4.66.1
pyyaml==6.0.1
